requirements.txt
aniso8601==9.0.1
bcrypt==4.0.0
click==8.1.3
colorama==0.4.5
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-JWT==0.3.2
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
PyJWT==1.4.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==1.4.41
typing_extensions==4.3.0
Werkzeug==2.2.2
zipp==3.8.1
